requirements.txt
# Core automation dependencies
selenium==4.15.0
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
psutil==5.9.5

# DevOps integrations
kubernetes==28.1.0
boto3==1.29.0  # AWS SDK
azure-mgmt-compute==30.0.0  # Azure SDK
prometheus-client==0.18.0
docker==6.1.3

# Data processing
pandas==2.1.0
numpy==1.24.3

# Web framework (for API endpoints)
flask==2.3.3
flask-cors==4.0.0

# Logging and monitoring
loguru==0.7.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0
